one-more-burger
- stworzenie i uzycie kontektsu
- usuniecie cartData
