--- src/Summary/cartData.js
@@ -1,19 +1,0 @@
-const cartData = [
-  {
-    name: "Cheeseburger",
-    price: 37,
-    quantity: 2,
-  },
-  {
-    name: "Bacon Burger",
-    price: 44,
-    quantity: 3,
-  },
-  {
-    name: "Crispy Chicken",
-    price: 41,
-    quantity: 1,
-  },
-];
-
-export default cartData;
